Move mid session token logic to app route
